requirements.txt
fastapi
python-dotenv
PyPDF2
langchain
uvicorn
pinecone-client
langchain-chroma
langchain-community
openai